+++ system-tests/debug-logging-test/src/main/java/systemundertest/SystemUnderTest.java
@@ -1,0 +1,10 @@
+package systemundertest;
+
+/** Fake system under test to generate some coverage. */
+public class SystemUnderTest {
+
+	public int foo() {
+		return 2;
+	}
+
+}