+++ CHANGELOG.md
@@ -8,6 +8,9 @@ We use [semantic versioning](http://semver.org/):
 
 - [feature] add Maven plugin for TIA
 - [feature] added official support for Java 17 and 18 and experimental support for Java 19.
+- [feature] New command line option: `debug`. Simplifies debugging by avoiding the process of providing an XML logging
+  configuration file. In debug mode logs are written to console and a configurable directory. For more details, see
+  [here](agent/README.md#general-options).
 - [breaking change] Artifactory uploader: use of new default upload schema for easy integration with Teamscale. To keep the old behavior add the `artifactory-legacy-path=true` option. For more details see 
   [here](agent/README.md#options-for-the-artifactory-upload).