+++ CHANGELOG.md
@@ -11,6 +11,8 @@ We use [semantic versioning](http://semver.org/):
 - [feature] New command line option: `debug`. Simplifies debugging by avoiding the process of providing an XML logging
   configuration file. In debug mode logs are written to console and a configurable directory. For more details, see
   [here](agent/README.md#general-options).
+- [breaking change] Artifactory uploader: use of new default upload schema for easy integration with Teamscale. To keep the old behavior add the `artifactory-legacy-path=true` option. For more details see 
+  [here](agent/README.md#options-for-the-artifactory-upload).
 
 # 23.1.1