+++ CHANGELOG.md
@@ -6,6 +6,8 @@ We use [semantic versioning](http://semver.org/):
 
 # next release
 
+# 24.0.0
+
 - [feature] add Maven plugin for TIA
 - [feature] added official support for Java 17 and 18 and experimental support for Java 19.
 - [feature] New command line option: `debug`. Simplifies debugging by avoiding the process of providing an XML logging